+++ src/App.spec.tsx
@@ -4,7 +4,7 @@ import { describe, it, expect, beforeEach, afterEach, vi } from 'vitest'
 import { render, screen, cleanup } from '@testing-library/react'
 import userEvent, { UserEvent } from '@testing-library/user-event'
 import { PictRunner } from '@takeyaqa/pict-browser'
-import App from '../src/App'
+import App from './App'
 
 describe('App', () => {
   describe('ParametersArea', () => {
