+++ src/features/menu/components/index.ts
@@ -1,0 +1,1 @@
+export { default as MenuSection } from './MenuSection'
