+++ src/features/model/components/index.ts
@@ -1,0 +1,3 @@
+export { default as ConstraintsSection } from './ConstraintsSection'
+export { default as ParametersSection } from './ParametersSection'
+export { default as SubModelsSection } from './SubModelsSection'
