+++ src/features/result/components/index.ts
@@ -1,0 +1,1 @@
+export { default as ResultSection } from './ResultSection'
