--- src/sections/index.ts
@@ -1,9 +1,0 @@
-export { default as HeaderSection } from './HeaderSection'
-export { default as NotificationMessageSection } from './NotificationMessageSection'
-export { default as MenuSection } from './MenuSection'
-export { default as ParametersSection } from './ParametersSection'
-export { default as OptionsSection } from './OptionsSection'
-export { default as SubModelsSection } from './SubModelsSection'
-export { default as ConstraintsSection } from './ConstraintsSection'
-export { default as ResultSection } from './ResultSection'
-export { default as FooterSection } from './FooterSection'
