+++ src/shared/components/Analytics.tsx
