+++ src/shared/components/FooterSection.tsx
