+++ src/shared/components/NotificationMessageSection.tsx
