+++ src/shared/components/Section.tsx
