+++ src/shared/components/index.ts
@@ -3,6 +3,9 @@ export { default as Analytics } from './Analytics'
 export { default as AugmentDiv } from './AugmentDiv'
 export { default as Button } from './Button'
 export { default as Checkbox } from './Checkbox'
+export { default as FooterSection } from './FooterSection'
+export { default as HeaderSection } from './HeaderSection'
+export { default as NotificationMessageSection } from './NotificationMessageSection'
 export { default as NumberInput } from './NumberInput'
 export { default as Section } from './Section'
 export { default as Switch } from './Switch'
